VPP-939: Update CLI Helptext for rx-placement commands
The CLI commands were change from 'set/show interface placement' to
'set/show interface rx-placement', but the associated help text was not
updated. On the 'set interface rx-placement', some of the parameters
were renamed, which was confusing when looking at the documentation for
help.

While in the file interface_cli.c, cleaned up several other CLI
commands. There are still other commands in the file that need
addressing.

Change-Id: Ifdc357c73ad9f3362133c495217b5a6b3a411eab
Signed-off-by: Billy McFall <bmcfall@redhat.com>