null-terminate some formatted string
Any u8* variable created by format() is NOT null-terminated.
Add the null terminating byte with vec_terminate_c_string().

If that variable is used by (at least) hash_get_mem(), then it needs to
be null-terminated, as it will go through string_key_sum() which makes a
call to strlen.

Change-Id: I4e51e1b6668f557e53af3bb897cd281598eedbc0
Signed-off-by: Gabriel Ganne <gabriel.ganne@enea.com>