build: allow to redefine max_lcores for DPDK
In current versions of DPDK (at least as of 24.11), max
amount of cores available determined at compile_time.

Unless machine type 'native' is used, it is statically set
per platform and for most of them equals to 128.

Add a new option to allow user to redefine max amount of
lcores available to DPDK.

Type: make

Change-Id: Ic4356e12ca75a5c84aeb3bee72de432a23aa6f8a
Signed-off-by: Vladimir Smirnov <civil.over@gmail.com>